+++ sparv/core/snake_utils.py
@@ -6,7 +6,7 @@ import re
 from collections import OrderedDict, defaultdict
 from itertools import combinations
 from pathlib import Path
-from typing import Any, List, Optional, Set, Tuple
+from typing import Any, List, Optional, Set, Tuple, Union
 
 import snakemake
 from snakemake.io import expand
@@ -370,9 +370,8 @@ def rule_helper(rule: RuleStorage, config: dict, storage: SnakeStorage, config_m
         elif param_type in (ExportAnnotations, ExportAnnotationNames, ExportAnnotationsAllSourceFiles):
             if not isinstance(param_value, param_type):
                 param_value = param_type(param_value)
-            rule.parameters[param_name] = param_value
 
-            source = param.default.config_name
+            source = param_value.config_name
             annotations = sparv_config.get(source, [])
             if not annotations:
                 rule.missing_config.add(source)
@@ -398,21 +397,26 @@ def rule_helper(rule: RuleStorage, config: dict, storage: SnakeStorage, config_m
                 if a not in plain_annotations:
                     export_annotations.append((annotation_type(a), None))
 
+            items = []
+
             for annotation, export_name in export_annotations:
-                if param.default.is_input:
+                if param_value.is_input:
                     if param_type == ExportAnnotationsAllSourceFiles:
                         rule.inputs.extend(
                             expand(escape_wildcards(paths.work_dir / get_annotation_path(annotation.name)),
                                    file=storage.source_files))
                     else:
                         rule.inputs.append(paths.work_dir / get_annotation_path(annotation.name))
-                rule.parameters[param_name].append((annotation, export_name))
+                items.append((annotation, export_name))
+            param_value.items = items
+            rule.parameters[param_name] = param_value
         # SourceAnnotations
         elif param_type in (SourceAnnotations, SourceAnnotationsAllSourceFiles):
-            rule.parameters[param_name] = param_type(
-                param.default.config_name,
-                sparv_config.get(param.default.config_name)
-            )
+            if not isinstance(param_value, param_type):
+                param_value = param_type(param_value)
+            param_value: Union[SourceAnnotations, SourceAnnotationsAllSourceFiles]
+            param_value.raw_list = sparv_config.get(param_value.config_name)
+            rule.parameters[param_name] = param_value
             if param_type == SourceAnnotationsAllSourceFiles:
                 rule.parameters[param_name].source_files = storage.source_files
                 rule.inputs.extend(
@@ -422,24 +426,25 @@ def rule_helper(rule: RuleStorage, config: dict, storage: SnakeStorage, config_m
                 rule.inputs.append(paths.work_dir / get_annotation_path(io.STRUCTURE_FILE, data=True))
         # HeaderAnnotations
         elif param_type in (HeaderAnnotations, HeaderAnnotationsAllSourceFiles):
-            rule.parameters[param_name] = param_type(
-                param.default.config_name,
-                sparv_config.get(param.default.config_name)
-            )
+            if not isinstance(param_value, param_type):
+                param_value = param_type(param_value)
+            param_value.raw_list = sparv_config.get(param_value.config_name)
+            rule.parameters[param_name] = param_value
             if param_type == HeaderAnnotationsAllSourceFiles:
                 rule.parameters[param_name].source_files = storage.source_files
         # Corpus
-        elif param.annotation == Corpus:
+        elif param_type == Corpus:
             rule.parameters[param_name] = Corpus(sparv_config.get("metadata.id"))
         # Language
-        elif param.annotation == Language:
+        elif param_type == Language:
             rule.parameters[param_name] = Language(sparv_config.get("metadata.language"))
         # SourceFilename
-        elif param.annotation == SourceFilename:
+        elif param_type == SourceFilename:
             rule.file_parameters.append(param_name)
         # AllSourceFilenames (all source filenames)
         elif param_type == AllSourceFilenames:
-            rule.parameters[param_name] = AllSourceFilenames(storage.source_files)
+            param_value.items = storage.source_files
+            rule.parameters[param_name] = param_value
         # Text
         elif param_type == Text:
             text_path = Path("{file}") / io.TEXT_FILE
@@ -467,22 +472,22 @@ def rule_helper(rule: RuleStorage, config: dict, storage: SnakeStorage, config_m
                 else:
                     rule.parameters[param_name] = model_param[0]
         # Binary
-        elif param.annotation in (Binary, BinaryDir):
+        elif param_type in (Binary, BinaryDir):
             rule.configs.update(registry.find_config_variables(param.default))
             rule.classes.update(registry.find_classes(param.default))
             param_value, missing_configs = registry.expand_variables(param.default, rule.full_name)
             rule.missing_config.update(missing_configs)
-            binary = util.system.find_binary(param_value, executable=False, allow_dir=param.annotation == BinaryDir)
+            binary = util.system.find_binary(param_value, executable=False, allow_dir=param_type == BinaryDir)
             if not binary:
                 rule.missing_binaries.add(param_value)
             binary = Path(binary if binary else param_value)
             rule.inputs.append(binary)
-            rule.parameters[param_name] = param.annotation(binary)
+            rule.parameters[param_name] = param_type(binary)
         # Source
-        elif param.annotation == Source:
+        elif param_type == Source:
             rule.parameters[param_name] = Source(get_source_path())
         # Export
-        elif param.annotation == Export:
+        elif param_type == Export:
             rule.configs.update(registry.find_config_variables(param.default))
             rule.classes.update(registry.find_classes(param.default))
             param_value, missing_configs = registry.expand_variables(param.default, rule.full_name)
@@ -496,7 +501,7 @@ def rule_helper(rule: RuleStorage, config: dict, storage: SnakeStorage, config_m
             if "{" in param_value:
                 rule.wildcard_annotations.append(param_name)
         # ExportInput
-        elif param.annotation == ExportInput:
+        elif param_type == ExportInput:
             rule.configs.update(registry.find_config_variables(param.default))
             rule.classes.update(registry.find_classes(param.default))
             param_value, missing_configs = registry.expand_variables(param.default, rule.full_name)