+++ CHANGELOG.md
@@ -41,8 +41,10 @@
 - `korp:relations`, `korp:relations_sql` and `korp:install_relations` has been renamed to `korp:wordpicture`,
   `korp:wordpicture_sql` and `korp:install_wordpicture` respectively.
 - Target path is no longer optional for the utility functions `install_path` and `rsync`.
-- The classes `SourceAnnotations` and `SourceAnnotationsAllSourceFiles` are now pre-parsed iterables instead of lists
-  that need parsing and expanding.
+- The classes `SourceAnnotations` and `SourceAnnotationsAllSourceFiles` are now pre-parsed, immutable iterables instead
+  of lists that need parsing and expanding.
+- The classes `AllSourceFilenames`, `ExportAnnotations`, `ExportAnnotationsAllSourceFiles` and `ExportAnnotationNames`
+  are now immutable iterables instead of lists.
 
 ### Fixed